--- Ch2/helloworld_finished.py
@@ -1,9 +1,0 @@
-#
-# Example file for HelloWorld
-#
-
-def main():
-  print ("hello world!")
-    
-if __name__ == "__main__":
-  main()
