--- Ch2/helloworld_start.py
@@ -1,9 +1,0 @@
-#
-# Example file for HelloWorld
-#
-
-def main():
-  print("Hello world")
-
-if __name__ == "__main__":
-    main()
