--- Ch3/calendars_start.py
@@ -1,26 +1,0 @@
-#
-# Example file for working with Calendars
-#
-
-# import the calendar module
-
-
-# create a plain text calendar
-
-
-# create an HTML formatted calendar
-
-
-# loop over the days of a month
-# zeroes mean that the day of the week is in an overlapping month
-
-  
-# The Calendar module provides useful utilities for the given locale,
-# such as the names of days and months in both full and abbreviated forms
-
-
-# Calculate days based on a rule: For example, consider
-# a team meeting on the first Friday of every month.
-# To figure out what days that would be for each month,
-# we can use this script:
-
