--- Ch4/ospathutils_start.py
@@ -1,29 +1,0 @@
-#
-# Example file for working with os.path module
-#
-import os
-from os import path
-import datetime
-from datetime import date, time, timedelta
-import time
-
-
-def main():
-  # Print the name of the OS
-
-
-  # Check for item existence and type
-
-  
-  # Work with file paths
-
-  
-  # Get the modification time
-
-  
-  # Calculate how long ago the item was modified
-
-
-  
-if __name__ == "__main__":
-  main()
