--- Ch4/shell_start.py
@@ -1,31 +1,0 @@
-#
-# Example file for working with filesystem shell methods
-#
-import os
-from os import path
-import shutil
-
-
-def main():
-  # make a duplicate of an existing file
-  if path.exists("textfile.txt"):
-    # get the path to the file in the current directory
-
-    
-    # let's make a backup copy by appending "bak" to the name
-
-    
-    # copy over the permissions, modification times, and other info
-
-
-    # rename the original file
-
-    
-    # now put things into a ZIP archive
-
-
-    # more fine-grained control over ZIP files
-
-      
-if __name__ == "__main__":
-  main()
