--- Ch5/htmlparsing_start.py
@@ -1,13 +1,0 @@
-# 
-# Example file for parsing and processing HTML
-#
-
-
-def main():
-  # instantiate the parser and feed it some HTML
-  parser = MyHTMLParser()
-    
-
-if __name__ == "__main__":
-  main();
-
