--- Ch5/inetdata_start.py
@@ -1,7 +1,0 @@
-# 
-# Example file for retrieving data from the internet
-#
-
-
-if __name__ == "__main__":
-  main()
