--- Ch5/samplehtml.html
@@ -1,35 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <title>Sample HTML Document</title>
-    <meta name="description" content="This is a sample HTML file" />
-    <meta name="author" content="Administrator" />
-    <meta name="viewport" content="width=device-width; initial-scale=1.0" />
-    <!-- Replace favicon.ico & apple-touch-icon.png in the root of your domain and delete these references -->
-    <link rel="shortcut icon" href="/favicon.ico" />
-    <link rel="apple-touch-icon" href="/apple-touch-icon.png" />
-  </head>
-
-  <body>
-    <div>
-      <header>
-        <h1>HTML Sample File</h1>
-      </header>
-      <nav>
-        <p>
-          <a href="/">Home</a>
-        </p>
-        <p>
-          <a href="/contact">Contact</a>
-        </p>
-      </nav>
-      <div>
-
-      </div>
-      <footer>
-        <p>&copy; Copyright by Administrator</p>
-      </footer>
-    </div>
-  </body>
-</html>
