--- Ch5/samplexml.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<person>
-  <firstname>Joe</firstname>
-  <lastname>Marini</lastname>
-  <home>Seattle</home>
-  <skill name="JavaScript"/>
-  <skill name="Python"/>
-  <skill name="C#"/>
-  <skill name="HTML"/>
-</person>
